ROC-7242 | renaming classes
