+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/caseworker/ReferToJudgeCallbackHandler.java
@@ -1,57 +1,74 @@
 package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.caseworker;
 
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import org.springframework.beans.factory.annotation.Value;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.cmc.ccd.domain.CCDCase;
+import uk.gov.hmcts.cmc.ccd.domain.CCDLaList;
 import uk.gov.hmcts.cmc.ccd.domain.CaseEvent;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.Role;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.Callback;
+import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackHandler;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackParams;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackType;
 import uk.gov.hmcts.cmc.claimstore.utils.CaseDetailsConverter;
-import uk.gov.hmcts.cmc.domain.models.ClaimState;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
 
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.REFER_TO_JUDGE_BY_CLAIMANT;
-import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.REFER_TO_JUDGE_BY_DEFENDANT;
-import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.CITIZEN;
+import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.REFER_TO_JUDGE;
+import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.CASEWORKER;
 
 @Service
-public class ReferToJudgeCallbackHandler extends AbstractStateChangeCallbackHandler {
-    private static final List<CaseEvent> EVENTS =
-        Arrays.asList(REFER_TO_JUDGE_BY_CLAIMANT, REFER_TO_JUDGE_BY_DEFENDANT);
-    private static final List<Role> ROLES = Collections.singletonList(CITIZEN);
+@ConditionalOnProperty("feature_toggles.ctsc_enabled")
+public class ReferToJudgeCallbackHandler extends CallbackHandler {
 
-    private final boolean ctscEnabled;
+    private static final List<CaseEvent> EVENTS = Collections.singletonList(REFER_TO_JUDGE);
+    private static final List<Role> ROLES = ImmutableList.of(CASEWORKER);
 
-    private final ImmutableMap<CallbackType, Callback> callbacks = ImmutableMap.of(
-        CallbackType.ABOUT_TO_SUBMIT, this::determineState
-    );
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+    private final CaseDetailsConverter caseDetailsConverter;
 
+    @Autowired
     public ReferToJudgeCallbackHandler(
-        CaseDetailsConverter caseDetailsConverter,
-        @Value("${feature_toggles.ctsc_enabled}") boolean ctscEnabled) {
-
-        super(EVENTS, ROLES, caseDetailsConverter);
-        this.ctscEnabled = ctscEnabled;
-    }
-
-    private CallbackResponse determineState(CallbackParams callbackParams) {
-        ClaimState state = ctscEnabled ? ClaimState.REDETERMINATION_REQUESTED : ClaimState.OPEN;
-        return AboutToStartOrSubmitCallbackResponse.builder()
-            .data(updateState(callbackParams, state))
-            .build();
+        CaseDetailsConverter caseDetailsConverter
+    ) {
+        this.caseDetailsConverter = caseDetailsConverter;
     }
 
     @Override
     protected Map<CallbackType, Callback> callbacks() {
-        return callbacks;
+        return ImmutableMap.of(CallbackType.ABOUT_TO_SUBMIT, this::assignToJudge);
     }
 
+    @Override
+    public List<CaseEvent> handledEvents() {
+        return EVENTS;
+    }
+
+    @Override
+    public List<Role> getSupportedRoles() {
+        return ROLES;
+    }
+
+    private CallbackResponse assignToJudge(CallbackParams callbackParams) {
+        CallbackRequest callbackRequest = callbackParams.getRequest();
+
+        CCDCase ccdCase = caseDetailsConverter.extractCCDCase(callbackRequest.getCaseDetails());
+        ccdCase.setAssignedTo(CCDLaList.REFER_TO_JUDGE);
+        Map<String, Object> dataMap = caseDetailsConverter.convertToMap(ccdCase);
+
+        return AboutToStartOrSubmitCallbackResponse
+            .builder()
+            .data(dataMap)
+            .build();
+    }
 }
