--- src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/caseworker/ReferToJudgeCallbackHandler1.java
@@ -1,74 +1,0 @@
-package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.caseworker;
-
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.stereotype.Service;
-import uk.gov.hmcts.cmc.ccd.domain.CCDCase;
-import uk.gov.hmcts.cmc.ccd.domain.CCDLaList;
-import uk.gov.hmcts.cmc.ccd.domain.CaseEvent;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.Role;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.Callback;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackHandler;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackParams;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackType;
-import uk.gov.hmcts.cmc.claimstore.utils.CaseDetailsConverter;
-import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-
-import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.REFER_TO_JUDGE;
-import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.CASEWORKER;
-
-@Service
-@ConditionalOnProperty("feature_toggles.ctsc_enabled")
-public class ReferToJudgeCallbackHandler1 extends CallbackHandler {
-
-    private static final List<CaseEvent> EVENTS = Collections.singletonList(REFER_TO_JUDGE);
-    private static final List<Role> ROLES = ImmutableList.of(CASEWORKER);
-
-    private final Logger logger = LoggerFactory.getLogger(getClass());
-    private final CaseDetailsConverter caseDetailsConverter;
-
-    @Autowired
-    public ReferToJudgeCallbackHandler1(
-        CaseDetailsConverter caseDetailsConverter
-    ) {
-        this.caseDetailsConverter = caseDetailsConverter;
-    }
-
-    @Override
-    protected Map<CallbackType, Callback> callbacks() {
-        return ImmutableMap.of(CallbackType.ABOUT_TO_SUBMIT, this::assignToJudge);
-    }
-
-    @Override
-    public List<CaseEvent> handledEvents() {
-        return EVENTS;
-    }
-
-    @Override
-    public List<Role> getSupportedRoles() {
-        return ROLES;
-    }
-
-    private CallbackResponse assignToJudge(CallbackParams callbackParams) {
-        CallbackRequest callbackRequest = callbackParams.getRequest();
-
-        CCDCase ccdCase = caseDetailsConverter.extractCCDCase(callbackRequest.getCaseDetails());
-        ccdCase.setAssignedTo(CCDLaList.REFER_TO_JUDGE);
-        Map<String, Object> dataMap = caseDetailsConverter.convertToMap(ccdCase);
-
-        return AboutToStartOrSubmitCallbackResponse
-            .builder()
-            .data(dataMap)
-            .build();
-    }
-}
